+++ compiler/src/package.rs
@@ -6,6 +6,8 @@ use self::error::{Error, Result};
 use self::schema::Schema;
 use self::spec::Spec;
 use flate2::read::GzDecoder;
+use flate2::write::GzEncoder;
+use flate2::Compression;
 use std::fs::File;
 use std::path::PathBuf;
 use tar::Archive;
@@ -13,6 +15,7 @@ use tempfile::TempDir;
 
 const SCHEMA_FILE: &str = "values.schema.json";
 const SPEC_FILE: &str = "kcp.json";
+const ARCHIVE_EXTENSION: &str = "tgz";
 
 pub struct Package {
 	pub root: PathBuf,
@@ -20,6 +23,7 @@ pub struct Package {
 	pub schema: Option<Schema>,
 }
 
+/// Associated functions
 impl Package {
 	pub fn from_path(root: PathBuf) -> Result<Self> {
 		let root = match root.extension() {
@@ -43,9 +47,27 @@ impl Package {
 	}
 }
 
+/// Methods
+impl Package {
+	pub fn compress(self, dir: &PathBuf) -> std::result::Result<PathBuf, String> {
+		let mut target = dir.clone();
+		target.push(self.spec.name);
+		target.set_extension(ARCHIVE_EXTENSION);
+
+		let file = File::create(target.clone()).map_err(|err| err.to_string())?;
+		let enc = GzEncoder::new(file, Compression::default());
+		let mut tar = tar::Builder::new(enc);
+
+		tar.append_dir_all("", self.root)
+			.map_err(|err| err.to_string())?;
+
+		Ok(target)
+	}
+}
+
 fn decompress(archive: PathBuf) -> Result<PathBuf> {
 	let ext = archive.extension().unwrap().to_str();
-	if ext != Some("tgz") {
+	if ext != Some(ARCHIVE_EXTENSION) {
 		return Err(Error::InvalidFormat);
 	}