+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - compile KCPs from `.tgz` archives with files at root
+- `package` command to create `.tgz` archives for valid KCPs
 
 ## [0.1.0] - 2020-09-29